added final changes
