+++ src/pages/homeCollection/allParticipants.js
@@ -1,6 +1,5 @@
-import { showAnimation, hideAnimation } from "../../shared.js";
+import { showAnimation, hideAnimation, getParticipantSelection } from "../../shared.js";
 import { renderParticipantSelectionHeader } from "./participantSelectionHeaders.js";
-import { getParticipantSelection } from "../../utils.js";
 import { participantSelectionDropdown } from "./printAddresses.js";
 import { nonUserNavBar, unAuthorizedUser } from "./../../navbar.js";
