+++ src/shared.js
@@ -858,3 +858,16 @@ export const SSOConfig = (inputValue) => {
     };
     return {tenantID, provider}
 }
+
+export const getParticipantSelection = async (filter) => {
+    const idToken = await getIdToken();
+    const response = await fetch(`https://us-central1-nih-nci-dceg-connect-dev.cloudfunctions.net/biospecimen?api=getParticipantSelection&type=${filter}`, 
+    {
+      method: "GET",
+      headers: {
+        Authorization: "Bearer " + idToken,
+      },
+    });
+    return response.json();
+  }
+
